price_level restoran ve kafeler harici silindi, route generation 1bug düzeltildi, explore page resize lag düzeltildi, cluster içindeki sayı rengi değiştirildi

+++ backend/src/main/java/com/roadrunner/place/loader/SeparatedPlaceDataLoader.java
@@ -237,7 +237,7 @@ public class SeparatedPlaceDataLoader implements ApplicationRunner {
     private Integer parseInt(String s) {
         if (s == null || s.isBlank()) return null;
         try {
-            return Integer.parseInt(s.strip());
+            return (int) Double.parseDouble(s.strip());
         } catch (NumberFormatException e) {
             return null;
         }
